[not-tested-yet] Refactor eosgrpc to use the same driver code as eos
