+++ pkg/storage/fs/eoshome/eoshome.go
@@ -19,9 +19,6 @@
 package eoshome
 
 import (
-	"bytes"
-	"encoding/gob"
-
 	"github.com/cs3org/reva/pkg/storage"
 	"github.com/cs3org/reva/pkg/storage/fs/registry"
 	"github.com/cs3org/reva/pkg/storage/utils/eosfs"
@@ -33,79 +30,8 @@ func init() {
 	registry.Register("eoshome", New)
 }
 
-type config struct {
-	// Namespace for metadata operations
-	Namespace string `mapstructure:"namespace" docs:"/"`
-
-	// ShadowNamespace for storing shadow data
-	ShadowNamespace string `mapstructure:"shadow_namespace" docs:"/.shadow"`
-
-	// UploadsNamespace for storing upload data
-	UploadsNamespace string `mapstructure:"uploads_namespace" docs:"/.uploads"`
-
-	// ShareFolder defines the name of the folder in the
-	// shadowed namespace. Ex: /eos/user/.shadow/h/hugo/MyShares
-	ShareFolder string `mapstructure:"share_folder" docs:"/MyShares"`
-
-	// Location of the eos binary.
-	// Default is /usr/bin/eos.
-	EosBinary string `mapstructure:"eos_binary" docs:"/usr/bin/eos"`
-
-	// Location of the xrdcopy binary.
-	// Default is /usr/bin/xrdcopy.
-	XrdcopyBinary string `mapstructure:"xrdcopy_binary" docs:"/usr/bin/xrdcopy"`
-
-	// URL of the Master EOS MGM.
-	// Default is root://eos-example.org
-	MasterURL string `mapstructure:"master_url" docs:"root://eos-example.org"`
-
-	// URL of the Slave EOS MGM.
-	// Default is root://eos-example.org
-	SlaveURL string `mapstructure:"slave_url" docs:"root://eos-example.org"`
-
-	// Location on the local fs where to store reads.
-	// Defaults to os.TempDir()
-	CacheDirectory string `mapstructure:"cache_directory" docs:"/var/tmp/"`
-
-	// SecProtocol specifies the xrootd security protocol to use between the server and EOS.
-	SecProtocol string `mapstructure:"sec_protocol" docs:"-"`
-
-	// Keytab specifies the location of the keytab to use to authenticate to EOS.
-	Keytab string `mapstructure:"keytab" docs:"-"`
-
-	// SingleUsername is the username to use when SingleUserMode is enabled
-	SingleUsername string `mapstructure:"single_username" docs:"-"`
-
-	// UserLayout wraps the internal path with user information.
-	// Example: if conf.Namespace is /eos/user and received path is /docs
-	// and the UserLayout is {{.Username}} the internal path will be:
-	// /eos/user/<username>/docs
-	UserLayout string `mapstructure:"user_layout" docs:"{{.Username}}"`
-
-	// Enables logging of the commands executed
-	// Defaults to false
-	EnableLogging bool `mapstructure:"enable_logging" docs:"false"`
-
-	// ShowHiddenSysFiles shows internal EOS files like
-	// .sys.v# and .sys.a# files.
-	ShowHiddenSysFiles bool `mapstructure:"show_hidden_sys_files" docs:"false"`
-
-	// ForceSingleUserMode will force connections to EOS to use SingleUsername
-	ForceSingleUserMode bool `mapstructure:"force_single_user_mode" docs:"false"`
-
-	// UseKeyTabAuth changes will authenticate requests by using an EOS keytab.
-	UseKeytab bool `mapstructure:"use_keytab" docs:"false"`
-
-	// Whether to maintain the same inode across various versions of a file.
-	// Requires extra metadata operations if set to true
-	VersionInvariant bool `mapstructure:"version_invariant" docs:"true"`
-
-	// GatewaySvc stores the endpoint at which the GRPC gateway is exposed.
-	GatewaySvc string `mapstructure:"gatewaysvc" docs:"0.0.0.0:19000"`
-}
-
-func parseConfig(m map[string]interface{}) (*config, error) {
-	c := &config{}
+func parseConfig(m map[string]interface{}) (*eosfs.Config, error) {
+	c := &eosfs.Config{}
 	if err := mapstructure.Decode(m, c); err != nil {
 		err = errors.Wrap(err, "error decoding conf")
 		return nil, err
@@ -125,18 +51,7 @@ func New(m map[string]interface{}) (storage.FS, error) {
 	if err != nil {
 		return nil, err
 	}
+	c.EnableHome = true
 
-	var buf bytes.Buffer
-	err = gob.NewEncoder(&buf).Encode(&c)
-	if err != nil {
-		return nil, err
-	}
-	var conf eosfs.Config
-	err = gob.NewDecoder(&buf).Decode(&conf)
-	if err != nil {
-		return nil, err
-	}
-	conf.EnableHome = true
-
-	return eosfs.NewEOSFS(&conf)
+	return eosfs.NewEOSFS(c)
 }
